backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "beautifulsoup4",
    "playwright",
    "python-dotenv",
    "pydantic",
    "google-generativeai",
    "jinja2"
]
